Remove course_codes and course_choices fields from CourseRequirement model
