+++ src/models/courseRequirementModel.js
@@ -1,4 +1,3 @@
-import { CourseInfo } from "./courseInfoModel";
 import { ProgramInfo } from "./ProgramInfoModel";
 
 var REQUIREMENT_ID = 10000;
@@ -29,37 +28,33 @@ export const YEAR_TO_REQ_SECTION_MAP = {
 export class CourseRequirement {
     constructor(data) {
         this.number_of_courses = data && data.number_of_courses? data.number_of_courses : 0;
-        this.course_choices = data && data.course_choices ? data.course_choices : [];
-        for (var i in this.course_choices) {
-            // eslint-disable-next-line valid-typeof
-            if (typeof this.course_choices[i] !== "CourseInfo") {
-                this.course_choices[i] = new CourseInfo({...this.course_choices[i]})
-            }
+        if (data && data.course_codes_raw) {
+            this.course_codes_raw = data.course_codes_raw;
+        } else if (data && data.course_codes) {
+            this.course_codes_raw = data.course_codes;
+        } else {
+            this.course_codes_raw = "";
         }
-        this.course_codes = this.course_choices.map(choice => {
-            return choice.course_code
-        });
-        this.course_codes_raw = data && data.course_codes ? data.course_codes : "";
-        if (data && data.course_choices && data.course_choices.length === 1) this.selected_course = data.course_choices[0];
-        else if (data && data.selected_course) this.selected_course = data.selected_course;
+        this.number_of_choices = data && data.course_codes_raw ? data.course_codes_raw.split(/,\s|\sor\s|,/).length : 0;
+        if (data && data.selected_course) this.selected_course = data.selected_course;
         else this.selected_course = {course_code: "WAITING", course_number: 42};
 
         this.major = data && data.major ? data.major : [];
-        for (i in this.major) {
+        for (let i in this.major) {
             // eslint-disable-next-line valid-typeof
             if (typeof this.major[i] !== "ProgramInfo") {
                 this.major[i] = new ProgramInfo({...this.major[i]})
             }
         }
         this.minor = data && data.minor ? data.minor : [];
-        for (i in this.minor) {
+        for (let i in this.minor) {
             // eslint-disable-next-line valid-typeof
             if (typeof this.minor[i] !== "ProgramInfo") {
                 this.minor[i] = new ProgramInfo({...this.minor[i]})
             }
         }
         this.specialization = data && data.specialization ? data.specialization : [];
-        for (i in this.specialization) {
+        for (let i in this.specialization) {
             // eslint-disable-next-line valid-typeof
             if (typeof this.specialization[i] !== "ProgramInfo") {
                 this.specialization[i] = new ProgramInfo({...this.specialization[i]})
@@ -91,8 +86,7 @@ export class CourseRequirement {
         else if (data.additional_requirements === "4A") this.year = 15;
         else if (data.additional_requirements === "4B") this.year = 16;
         else{
-            this.year = this.course_choices.length ? this.course_choices[0].year : -1;
-            for(let course of this.course_choices) if (course.year !== this.year) this.year = -1;
+            this.year = this.selected_course && this.selected_course.course_code !== "WAITING" ? this.selected_course.year : -1;
         }
 
         this.clickedDelete = false;
@@ -100,7 +94,7 @@ export class CourseRequirement {
     }
 
     deselect() {
-        if (this.course_choices.length === 1) return;
+        if (this.number_of_choices === 1) return;
         this.selected_course = {course_code: "WAITING", course_number: 42}
     }
