+++ src/store/modules/programInfo.js
@@ -2,10 +2,10 @@ import axios from "axios";
 import { ProgramInfo } from "../../models/ProgramInfoModel"
 
 // Production Kubernetes API
-const backend_api = "";
+// const backend_api = "";
 
 // Dev API
-// const backend_api = "http://127.0.0.1:8000";
+const backend_api = "http://127.0.0.1:8000";
 
 //need to move the routes to the configs
 const state = {
